rename quick action to action button
