+++ src/models/index.ts
@@ -1,3 +1,4 @@
+export * from './actionButton.model';
 export * from './apiConfiguration.model';
 export * from './application.model';
 export * from './channel.model';
@@ -11,7 +12,6 @@ export * from './permission.model';
 export * from './positionAttribute.model';
 export * from './positionAttributeCategory.model';
 export * from './pullJob.model';
-export * from './quickActions.model';
 export * from './record.model';
 export * from './referenceData.model';
 export * from './resource.model';
